refactor: 검증 로직 도메인으로 위임
